Completely refactor the instruction definition code.
Instructions are now defined in integer.ss (this might change to general.ss or similar; float.ss and so on will follow). Instructions are in a fairly declarative form so the underlying representation can be changed at any time. At the moment instructions write immediately to a byte output port

Instruction.ss defines the supporting code for defining instructions

Reference.ss defines a datatype for memory references

Assemble.ss and test.ss have minimal changes to adapt to the new style.

Next up: better testing and more instruction definitions.

Note that opcode.ss is kept around for reference. It will be removed at a later date.